+++ client/src/Home/Home.tsx
@@ -1,4 +1,4 @@
-import { MealsWrapper as Meals } from '../Meals/src/Meals';
+import { MealsWrapper as Meals } from '../Meals/Meals';
 
 export const Home: React.FC = () => {
   return <Meals />;
